requirements.txt
discord.py==2.2.2
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
youtube-dl-server==0.3
youtube-dl==2020.07.28
beautifulsoup4==4.8.2
lxml==4.4.1
requests==2.23.0
requests-html==0.10.0
psycopg2-binary==2.8.5
urllib3==1.25.8
google-api-python-client==2.107.0
google-api-python-client-stubs==1.18.0 
spotipy==2.13.0
python-dotenv==1.0.0
pycord==0.1.1
